Create Modify ReefAngel App for IOS
From AI to PWMA PWMD:
I'll I did here was change what ORP and Salinity values from RAparams to get the PWMA and PWMD. 
Like I said through PM on RA forum...this is for personal use and didn't want to take the time to rewrite all the code to make it proper..
But as it stands now the PWMA and PWMD replace those ORP and Salinity values. I know it's not proper..but it was an easy fix. The iphone simulator works just like I would like it to work on my phone. 